버그 수정

+++ src/user/User.java
@@ -18,7 +18,7 @@ public class User {
     // 가입일자
     private LocalDate nowDate;
     // 잔액 -
-    private int myMoney;
+    private long myMoney;
     // 비밀번호 힌트 - 보물 1호
     private String myTreasure;
 
@@ -80,11 +80,11 @@ public class User {
         this.nowDate = nowDate;
     }
 
-    public int getMyMoney() {
+    public long getMyMoney() {
         return myMoney;
     }
 
-    public void setMyMoney(int myMoney) {
+    public void setMyMoney(long myMoney) {
         this.myMoney = myMoney;
     }
